% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcp-data.R
\docType{data}
\name{coriell}
\alias{coriell}
\title{Array CGH data set of Coriell cell lines}
\format{A data frame containing five variables: first is clone name, second is clone chromosome, third is   clone position, fourth and fifth  are log2ratio for two cell lines.}
\source{
\url{ https://www.nature.com/articles/ng754}
}
\usage{
coriell
}
\description{
These are two data array CGH studies of Coriell cell lines taken from the reference below.
}
\examples{
demo(coriell)

}
\references{
\enumerate{ 
\item Olshen, A. B., Venkatraman, E. S., Lucito, R., Wigler, M. (2004), Circular binary segmentation for the analysis of array-based DNA copy number data, \emph{Biostatistics}, \bold{5}, 557-572. url: \url{https://www.bioconductor.org/packages/release/bioc/html/DNAcopy.html}.
\item Snijders \emph{et al.} (2001), Assembly of microarrays for genome-wide measurement of DNA copy number, \emph{Nature Genetics}, \bold{29}, 263-264. 
}
}
\keyword{datasets}
