% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcp.R
\name{create_table_structure}
\alias{create_table_structure}
\title{Creates SQL table from R data frame structure}
\usage{
create_table_structure(conn, table_name, data, dialect = "t-sql")
}
\arguments{
\item{table_name}{SQL table name to be created}

\item{data}{Data frame to be used as prototype for SQL table creation}

\item{dialect}{SQL dialect for CREATE TABLE statement. As of now only T-SQL supported.}
}
\description{
Creates SQL table from R data frame structure
}
