% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcp.R
\name{bcp}
\alias{bcp}
\title{Bulk Upload Data Frame into SQL Table}
\usage{
bcp(
  conn,
  data,
  table_name,
  chunk_size = 5000L,
  show_progress = TRUE,
  auto_create_table = FALSE,
  stop_if_error = TRUE,
  drop_if_exists = FALSE
)
}
\arguments{
\item{conn}{RODBC connection return by odbcDriverConnect() function}

\item{data}{Data frame to be uploaded}

\item{table_name}{SQL table name, it must exist in database}

\item{chunk_size}{How many rows to load in each chunk}

\item{show_progress}{Logical flag which indicates whether to show data loading progress}

\item{auto_create_table}{Logical flag whether automatically create table structure before data uploading}

\item{stop_if_error}{Logical flag whether raise error if BCP operation failed or just return error code}
}
\description{
Bulk Upload Data Frame into SQL Table
}
