% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{fitted.bcp}
\alias{fitted.bcp}
\title{Extract model fitted values}
\usage{
\method{fitted}{bcp}(object, ...)
}
\arguments{
\item{object}{the result of a call to \code{bcp()}.}

\item{...}{(optional) additional arguments, ignored.}
}
\value{
Fitted values extracted from the \code{bcp} object.
}
\description{
fitted method for class \code{bcp}.
}
\examples{
##### A random sample from a few normal distributions #####
testdata <- c(rnorm(50), rnorm(50, 5, 1), rnorm(50))
bcp.0 <- bcp(testdata)
residuals(bcp.0)

}
\seealso{
\code{\link{plot.bcp}}, \code{\link{summary.bcp}}, and \code{\link{print.bcp}} for summaries of the results.
}
\author{
Xiaofei Wang, Chandra Erdman and John W. Emerson
}
\keyword{datasets}
