% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcp-data.R
\docType{data}
\name{NewHavenHousing}
\alias{NewHavenHousing}
\title{New Haven housing data}
\format{A matrix containing location, 2011 assessed value, and physical characteristics in the columns.}
\source{
The data can be scraped from the New Haven, CT Online Assessment Database \url{http://gis.vgsi.com/newhavenct/}
}
\usage{
NewHavenHousing
}
\description{
Location, 2011 assessed value, and physical characteristics for 244 houses in a region of New Haven, CT.
}
\examples{
 
\dontrun{
demo("NewHaven")
}

}
\references{
Xiaofei Wang and John W. Emerson (2015). Bayesian Change Point Analysis of Linear Models on General Graphs, \emph{Working Paper}.
}
\keyword{datasets}
