% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcp-data.R
\docType{data}
\name{RealInt}
\alias{RealInt}
\title{US Ex-post Real Interest Rate data, 1961(1):1986(3)}
\format{A quarterly time series from 1961(1) to 1986(3).}
\source{
The data is available online in the data archive of the
\emph{Journal of Applied Econometrics}. url: 
 \url{http://qed.econ.queensu.ca/jae/2003-v18.1/bai-perron/}.
}
\usage{
RealInt
}
\description{
US ex-post real interest rate: the three-month treasury bill
deflated by the CPI inflation rate.
}
\examples{
 
demo(RealInt)

}
\references{
\enumerate{ 
\item J. Bai and P. Perron (2003), Computation and Analysis of Multiple Structural Change Models, \emph{Journal of Applied Econometrics}, \bold{18}, 1-22. \url{http://qed.econ.queensu.ca/jae/2003-v18.1/bai-perron/}.
\item Achim Zeileis, Friedrich Leisch, Kurt Hornik, Christian Kleiber (2002), strucchange: An R Package for Testing for Structural Change in Linear Regression Models, \emph{Journal of Statistical Software}, \bold{7}(2), 1--38. \url{http://www.jstatsoft.org/v07/i02/}.
}
}
\keyword{datasets}
